Add an Easter Egg!
The egg is in the form of a function at this point, but will be able to
be activated via the console later.

It will run a while() loop for 100 seconds that will continuously put
up annoying alert() boxes the whole time so the browser won’t work
until it finishes.